Writing out columns for dplyr functions
It is not easy to predict what columns a user will submit with their input dataset. Since we have forced some column naming convention, we will take advantage of that. We know what the columns names should be and we can make sure we are only altering the ones we care about